financial-data-pipeline/scripts/transform_fx_rates.py
import psycopg2
import os
from dotenv import load_dotenv

load_dotenv()

DB_CONFIG = {
    "host":     os.getenv("DB_HOST"),
    "port":     os.getenv("DB_PORT"),
    "dbname":   os.getenv("DB_NAME"),
    "user":     os.getenv("DB_USER"),
    "password": os.getenv("DB_PASSWORD"),
}

def run_transformation():
    with open("sql/transform_fx_rates.sql", "r") as f:
        sql = f.read()

    conn = psycopg2.connect(**DB_CONFIG)
    cursor = conn.cursor()
    cursor.execute(sql)
    conn.commit()
    cursor.close()
    conn.close()
    print("✅ Transformation complete.")

if __name__ == "__main__":
    run_transformation()
