financial-data-pipeline/requirements.txt
requests
psycopg2-binary
python-dotenv
pandas
apache-airflow
